Adjust positions in JSON files and modify text in config
The position values in the Service Packages and Subscriptions JSON files were updated. The title and tagline in the docusaurus config file were also edited to better describe the Video Streaming Resell Platform as StreamWise vPaaS. Additionally, the positions of API Reference links in the navbar were changed from left to right.