Enhance Account Management docs with new sections and data
The updates improve the documentation for Account Management. An added section on Account Attributes and an added reference to the Users section further accentuates the account properties and user permissions details. It also emphasized HTTP methods in API instructions and appended information on API endpoints for user account management and editing an account. Thus, the documentation now gives more comprehensive guidance on using and managing accounts.