feat: Add support for customer metadata and settings
This commit adds support for customer metadata and settings. It includes updates to the `customers.md` file to provide information about the `metadata` and `customer_settings` properties on resources. Additionally, it informs about the overwrite behavior of these properties. The commit also includes recent user commits related to updating subscriptions information and adding the `subscription_metadata` property to resources.